Separed custom text form field into class and generator, adde text style mixin to generator
